// src/app/models/quest-to-create.ts
export interface QuestToCreate {
    category: string
    difficulty: string
    title: string
    question: string
    image: string
    video: string
    audio: string
    answer: string
    option1: string
    option2: string
    option3: string
    tipsAndLinks: string
    availabilityInDays: number
    }